更新配置文件路徑，升級 vllm 依賴版本，調整 VLLM 伺服器的數據類型，修正評估腳本中的環境變數，並更新 README 文件以反映 unbabel-comet 的安裝要求

+++ src/BenchWeaver/inference/vllm/server.py
@@ -44,7 +44,7 @@ class VLLMServer:
                 "serve", str(model_path),
                 "--enable-chunked-prefill", "False",
                 "--tensor-parallel-size", str(self.get_max_usable_devices()),
-                "--dtype", "half",
+                "--dtype", "bfloat16",
                 "--served-model-name", model_name,
                 "--disable-log-requests",
                 "--disable-log-stats",
